Safari-611.1.10/LayoutTests/js/dom/domjit-function-licm.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>

<div id="parentNode">
<div id="previousSibling"></div><div id="target"><div id="firstChild"></div><div id="lastChild"></div></div><div id="nextSibling"></div>
</div>

<script>
description('Test DOMJIT function will be LICM-ed.');

function test() {
    var div = document.createElement('div');
    var ret = "invalid";
    div.setAttribute("cocoa", "Cocoa");
    for (var i = 0; i < 1e4; ++i)
        ret = div.getAttribute("cocoa");
    return ret;
}
var result;
(function () {
    for (var i = 0; i < 100; ++i) {
        result = test();
        shouldBe(`result`, `"Cocoa"`, true);
    }
}());
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
